+++ src/pulumi/core-dtap/network/__init__.py
@@ -9,3 +9,4 @@ from . import routing
 from . import nsg
 from . import vnet
 from . import dns
+from . import privatelink_endpoints