+++ src/pulumi/core-dtap/network/dns.py
@@ -1,6 +1,7 @@
-import pulumi_azure_native as azure_native
+from pulumi_azure_native import network as net
 
 from project_config import platform_config
+from platform_shared import container_registry_private_endpoint_configs
 from management import resource_groups
 
 from .vnet import vnet
@@ -8,7 +9,7 @@ from .vnet import vnet
 # ----------------------------------------------------------------------------------------------------------------------
 # STORAGE BLOB PRIVATE DNS ZONE
 # ----------------------------------------------------------------------------------------------------------------------
-storage_blob_private_dns_zone = azure_native.network.PrivateZone(
+storage_blob_private_dns_zone = net.PrivateZone(
     resource_name="privatelink-blob-core-windows-net",
     location="Global",
     private_zone_name="privatelink.blob.core.windows.net",
@@ -16,7 +17,7 @@ storage_blob_private_dns_zone = azure_native.network.PrivateZone(
     tags=platform_config.tags,
 )
 
-storage_blob_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
+storage_blob_private_dns_zone_link = net.VirtualNetworkLink(
     resource_name="privatelink-blob-core-windows-net",
     virtual_network_link_name=vnet.name,
     location="Global",
@@ -24,7 +25,7 @@ storage_blob_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
     registration_enabled=False,
     resource_group_name=resource_groups["infra"].name,
     tags=platform_config.tags,
-    virtual_network=azure_native.network.SubResourceArgs(
+    virtual_network=net.SubResourceArgs(
         id=vnet.id,
     ),
 )
@@ -32,7 +33,7 @@ storage_blob_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
 # ----------------------------------------------------------------------------------------------------------------------
 # STORAGE DFS PRIVATE DNS ZONE
 # ----------------------------------------------------------------------------------------------------------------------
-storage_dfs_private_dns_zone = azure_native.network.PrivateZone(
+storage_dfs_private_dns_zone = net.PrivateZone(
     resource_name="privatelink-dfs-core-windows-net",
     location="Global",
     private_zone_name="privatelink.dfs.core.windows.net",
@@ -40,14 +41,14 @@ storage_dfs_private_dns_zone = azure_native.network.PrivateZone(
     tags=platform_config.tags,
 )
 
-storage_dfs_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
+storage_dfs_private_dns_zone_link = net.VirtualNetworkLink(
     resource_name="privatelink-dfs-core-windows-net",
     virtual_network_link_name=vnet.name,
     location="Global",
     private_zone_name=storage_dfs_private_dns_zone.name,
     registration_enabled=False,
     resource_group_name=resource_groups["infra"].name,
-    virtual_network=azure_native.network.SubResourceArgs(
+    virtual_network=net.SubResourceArgs(
         id=vnet.id,
     ),
     tags=platform_config.tags,
@@ -56,7 +57,7 @@ storage_dfs_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
 # ----------------------------------------------------------------------------------------------------------------------
 # KEYVAULT PRIVATE DNS ZONE
 # ----------------------------------------------------------------------------------------------------------------------
-key_vault_private_dns_zone = azure_native.network.PrivateZone(
+key_vault_private_dns_zone = net.PrivateZone(
     resource_name="privatelink-vaultcore-azure-net",
     location="Global",
     private_zone_name="privatelink.vaultcore.azure.net",
@@ -64,7 +65,7 @@ key_vault_private_dns_zone = azure_native.network.PrivateZone(
     tags=platform_config.tags,
 )
 
-key_vault_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
+key_vault_private_dns_zone_link = net.VirtualNetworkLink(
     resource_name="privatelink-vaultcore-azure-net",
     virtual_network_link_name=vnet.name,
     location="Global",
@@ -72,7 +73,43 @@ key_vault_private_dns_zone_link = azure_native.network.VirtualNetworkLink(
     registration_enabled=False,
     resource_group_name=resource_groups["infra"].name,
     tags=platform_config.tags,
-    virtual_network=azure_native.network.SubResourceArgs(
+    virtual_network=net.SubResourceArgs(
         id=vnet.id,
     ),
+)
+
+# ----------------------------------------------------------------------------------------------------------------------
+# CONTAINER REGISTRY PRIVATE DNS ZONE
+# ----------------------------------------------------------------------------------------------------------------------
+
+# The container registry DNS zone and zone link are deployed only if a container registry is also deployed in the shared
+# stack. This behavior might have to change if there is a need for container registry private link access that is not
+# dependent on the shared stack's registry.
+container_registry_dns_zone = (
+    net.PrivateZone(
+        resource_name="privatelink-azurecr-io",
+        location="Global",
+        private_zone_name="privatelink.azurecr.io",
+        resource_group_name=resource_groups["infra"].name,
+        tags=platform_config.tags,
+    )
+    if container_registry_private_endpoint_configs
+    else None
+)
+
+container_registry_dns_zone_link = (
+    net.VirtualNetworkLink(
+        resource_name="privatelink-azurecr-io",
+        virtual_network_link_name=vnet.name,
+        location="Global",
+        private_zone_name=container_registry_dns_zone.name,
+        registration_enabled=False,
+        resource_group_name=resource_groups["infra"].name,
+        tags=platform_config.tags,
+        virtual_network=net.SubResourceArgs(
+            id=vnet.id,
+        ),
+    )
+    if container_registry_private_endpoint_configs
+    else None
 )