+++ src/pulumi/core-shared/__main__.py
@@ -4,4 +4,5 @@ import analytics
 import automation
 import management
 import network
+import storage
 import security