+++ src/pulumi/core-shared/network/vnet.py
@@ -1,8 +1,9 @@
-import pulumi_azure_native as azure_native
+import pulumi_azure_native.network as net
+from pulumi import ResourceOptions
 from ingenii_azure_data_platform.utils import generate_cidr, generate_resource_name
+
 from management import resource_groups
 from project_config import platform_config, platform_outputs
-from pulumi import ResourceOptions
 
 from . import nat, routing
 
@@ -19,14 +20,12 @@ vnet_name = generate_resource_name(
     platform_config=platform_config,
 )
 
-vnet = azure_native.network.VirtualNetwork(
+vnet = net.VirtualNetwork(
     resource_name=vnet_name,
     virtual_network_name=vnet_name,
     resource_group_name=resource_groups["infra"].name,
     location=platform_config.region.long_name,
-    address_space=azure_native.network.AddressSpaceArgs(
-        address_prefixes=[vnet_address_space]
-    ),
+    address_space=net.AddressSpaceArgs(address_prefixes=[vnet_address_space]),
     tags=platform_config.tags,
     # Tags are added in the ignore_changes list because of:
     # https://github.com/ingenii-solutions/azure-data-platform/issues/71
@@ -53,7 +52,7 @@ outputs["virtual_network"] = {
 subnet_outputs = outputs["virtual_network"]["subnets"] = {}
 
 # GATEWAY SUBNET
-gateway_subnet = azure_native.network.Subnet(
+gateway_subnet = net.Subnet(
     resource_name=generate_resource_name(
         resource_type="subnet", resource_name="gateway", platform_config=platform_config
     ),
@@ -61,7 +60,7 @@ gateway_subnet = azure_native.network.Subnet(
     resource_group_name=resource_groups["infra"].name,
     virtual_network_name=vnet.name,
     address_prefix=generate_cidr(vnet_address_space, 24, 0),
-    route_table=azure_native.network.RouteTableArgs(id=routing.main_route_table.id),
+    route_table=net.RouteTableArgs(id=routing.main_route_table.id),
 )
 
 # Export subnet metadata
@@ -75,14 +74,14 @@ subnet_outputs["gateway"] = {
 privatelink_subnet_name = generate_resource_name(
     resource_type="subnet", resource_name="privatelink", platform_config=platform_config
 )
-privatelink_subnet = azure_native.network.Subnet(
+privatelink_subnet = net.Subnet(
     resource_name=privatelink_subnet_name,
     subnet_name=privatelink_subnet_name,
     resource_group_name=resource_groups["infra"].name,
     virtual_network_name=vnet.name,
     address_prefix=generate_cidr(vnet_address_space, 24, 1),
-    private_endpoint_network_policies=azure_native.network.VirtualNetworkPrivateEndpointNetworkPolicies.DISABLED,
-    route_table=azure_native.network.RouteTableArgs(id=routing.main_route_table.id),
+    private_endpoint_network_policies=net.VirtualNetworkPrivateEndpointNetworkPolicies.DISABLED,
+    route_table=net.RouteTableArgs(id=routing.main_route_table.id),
     opts=ResourceOptions(depends_on=[gateway_subnet]),
 )
 
@@ -99,24 +98,22 @@ hosted_services_subnet_name = generate_resource_name(
     resource_name="hosted-services",
     platform_config=platform_config,
 )
-hosted_services_subnet = azure_native.network.Subnet(
+hosted_services_subnet = net.Subnet(
     resource_name=hosted_services_subnet_name,
     subnet_name=hosted_services_subnet_name,
     resource_group_name=resource_groups["infra"].name,
     virtual_network_name=vnet.name,
     address_prefix=generate_cidr(vnet_address_space, 24, 2),
-    route_table=azure_native.network.RouteTableArgs(id=routing.main_route_table.id),
-    nat_gateway=azure_native.network.SubResourceArgs(id=nat.gateway.id),
+    route_table=net.RouteTableArgs(id=routing.main_route_table.id),
+    nat_gateway=net.SubResourceArgs(id=nat.gateway.id),
     service_endpoints=[
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.Storage",
-        ),
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.KeyVault",
-        ),
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.SQL",
-        ),
+        net.ServiceEndpointPropertiesFormatArgs(service=service)
+        for service in [
+            "Microsoft.ContainerRegistry",
+            "Microsoft.KeyVault",
+            "Microsoft.Storage",
+            "Microsoft.SQL",
+        ]
     ],
     opts=ResourceOptions(depends_on=[privatelink_subnet]),
 )
@@ -134,25 +131,23 @@ devops_deployment_subnet_name = generate_resource_name(
     resource_name="devops-deployment",
     platform_config=platform_config,
 )
-devops_deployment_subnet = azure_native.network.Subnet(
+devops_deployment_subnet = net.Subnet(
     resource_name=devops_deployment_subnet_name,
     subnet_name=devops_deployment_subnet_name,
     resource_group_name=resource_groups["infra"].name,
     virtual_network_name=vnet.name,
     address_prefix=generate_cidr(vnet_address_space, 24, 3),
-    route_table=azure_native.network.RouteTableArgs(id=routing.main_route_table.id),
+    route_table=net.RouteTableArgs(id=routing.main_route_table.id),
     service_endpoints=[
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.Storage",
-        ),
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.KeyVault",
-        ),
-        azure_native.network.ServiceEndpointPropertiesFormatArgs(
-            service="Microsoft.SQL",
-        ),
+        net.ServiceEndpointPropertiesFormatArgs(service=service)
+        for service in [
+            "Microsoft.ContainerRegistry",
+            "Microsoft.KeyVault",
+            "Microsoft.Storage",
+            "Microsoft.SQL",
+        ]
     ],
-    opts=ResourceOptions(depends_on=[privatelink_subnet]),
+    opts=ResourceOptions(depends_on=[hosted_services_subnet]),
 )
 
 # Export subnet metadata