+++ src/python/ingenii_azure_data_platform/ingenii_azure_data_platform/defaults.py
@@ -1,22 +1,21 @@
-from pulumi_azure_native.storage import (
-    NetworkRuleSetArgs as StorageNetworkRuleSetArgs,
-    DefaultAction as StorageNetworkRuleSetDefaultAction,
-)
+from pulumi_azure_native import storage, keyvault, containerregistry
 
-from pulumi_azure_native.keyvault import (
-    NetworkRuleSetArgs as KeyVaultNetworkRuleSetArgs,
-)
-
-KEY_VAULT_DEFAULT_FIREWALL = KeyVaultNetworkRuleSetArgs(
-    default_action="Allow",
-    bypass="AzureServices",
+STORAGE_ACCOUNT_DEFAULT_FIREWALL = storage.NetworkRuleSetArgs(
+    default_action=storage.DefaultAction.ALLOW,
+    bypass=storage.Bypass.AZURE_SERVICES,
     ip_rules=None,
     virtual_network_rules=None,
 )
 
-STORAGE_ACCOUNT_DEFAULT_FIREWALL = StorageNetworkRuleSetArgs(
-    default_action=StorageNetworkRuleSetDefaultAction("Allow"),
-    bypass="AzureServices",
+KEY_VAULT_DEFAULT_FIREWALL = keyvault.NetworkRuleSetArgs(
+    default_action=keyvault.NetworkRuleAction.ALLOW,
+    bypass=keyvault.NetworkRuleBypassOptions.AZURE_SERVICES,
     ip_rules=None,
     virtual_network_rules=None,
+)
+
+CONTAINER_REGISTRY_DEFAULT_FIREWALL = containerregistry.NetworkRuleSetArgs(
+    default_action=containerregistry.DefaultAction.ALLOW,
+    ip_rules=[],
+    virtual_network_rules=[],
 )